libcamera: software_isp: Use a specific integer type for black level
The documented range of values corresponds to uint8_t, so let's use that
type.

+++ src/ipa/simple/black_level.cpp
@@ -43,7 +43,7 @@ BlackLevel::BlackLevel()
  * \return The black level, in the range from 0 (minimum) to 255 (maximum).
  * If the black level couldn't be determined yet, return 0.
  */
-unsigned int BlackLevel::get() const
+uint8_t BlackLevel::get() const
 {
 	return blackLevelSet_ ? blackLevel_ : 0;
 }
